+++ src/tab_foundry/training/trainer.py
@@ -17,9 +17,9 @@ from torch.utils.data import DataLoader
 from omegaconf import DictConfig, OmegaConf
 
 from tab_foundry.data.factory import build_task_dataset, build_task_loader
+from tab_foundry.model.architectures.tabfoundry import ClassificationOutput, RegressionOutput
 from tab_foundry.model.factory import build_model_from_spec
 from tab_foundry.model.spec import model_build_spec_from_mappings
-from tab_foundry.model.architectures.tabiclv2 import ClassificationOutput, RegressionOutput
 from tab_foundry.types import TaskBatch, TrainResult
 
 from .batching import move_batch